requirements.txt
gql==2.0.0
python-dotenv==0.18.0